arkive-backend/users-service/requirements.txt
Flask
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy
flask-migrate
Jinja2==3.1.4
python-dotenv
SQLAlchemy==2.0.35
Werkzeug
flask-bcrypt
flask-jwt-extended
faker
psycopg2-binary